requirements.txt
pandas==1.2.2
requests==2.22.0
config==0.5.1
python-dotenv==0.19.2
